requirements.txt
# requirements.txt

streamlit
pandas
numpy
sentence-transformers
google-generativeai
pypdf